Add block elements Do not allows player walks over blocks

+++ pacman/src/main/java/pacman/Pacboy.java
@@ -3,6 +3,9 @@ package pacman;
 
 public class Pacboy implements Jogo {
 
+    public static final char CHAR_BLOCK = '|';
+    public static final char CHAR_FRUIT = '*';
+    public static final char CHAR_PLAYER = 'C';
     private char [][] state;
     private int size;
 
@@ -16,20 +19,29 @@ public class Pacboy implements Jogo {
 
         this.size = 5;
         this.state = new char[size][size];
-        for (int i = 0; i < this.size; i++) {
-            for (int j = 0; j < this.size; j++) {
-                this.state[i][j] = '*';
+        for (int y = 0; y < this.size; y++) {
+            for (int x = 0; x < this.size; x++) {
+                this.state[y][x] = CHAR_FRUIT;
+                if (shouldBeABlock(x, y)) {
+                    this.state[y][x] = CHAR_BLOCK;
+                }
             }
         }
 
-        this.state[pacboyX][pacboyY] = 'C';
+        this.state[pacboyX][pacboyY] = CHAR_PLAYER;
+    }
+
+    private boolean shouldBeABlock(int x, int y) {
+        return (x == 2 && y == 1) ||
+                (x == 3 && y == 2) ||
+                (x == 1 && y == 3);
     }
 
     public String tela() {
         StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < this.size; i++) {
-            for (int j = 0; j < this.size; j++) {
-                sb.append(this.state[i][j]);
+        for (int y = 0; y < this.size; y++) {
+            for (int x = 0; x < this.size; x++) {
+                sb.append(this.state[y][x]);
             }
             sb.append("\n");
         }
@@ -42,31 +54,51 @@ public class Pacboy implements Jogo {
     }
 
     public void direita() {
-        deletePacboy();
-        pacboyX = Math.abs((pacboyX + 1) % this.size);
+        int desiredX = Math.abs((pacboyX + 1) % this.size);
+        if (shouldBeABlock(desiredX, pacboyY)) {
+            return;
+        }
 
-        this.state[pacboyY][pacboyX] = 'C';
+        deletePacboy();
+        pacboyX = desiredX;
+
+        this.state[pacboyY][pacboyX] = CHAR_PLAYER;
     }
 
-    public void esquerda() {
-        deletePacboy();
-        pacboyX = Math.abs((pacboyX - 1) % this.size);
 
-        this.state[pacboyY][pacboyX] = 'C';
+    public void esquerda() {
+        int desiredX = Math.abs((pacboyX - 1) % this.size);
+        if (shouldBeABlock(desiredX, pacboyY)) {
+            return;
+        }
+
+        deletePacboy();
+        pacboyX = desiredX;
+
+        this.state[pacboyY][pacboyX] = CHAR_PLAYER;
     }
 
     public void sobe() {
-        deletePacboy();
-        pacboyY = Math.abs((pacboyY - 1) % this.size);
+        int desiredY = Math.abs((pacboyY - 1) % this.size);
+        if (shouldBeABlock(pacboyX, desiredY)) {
+            return;
+        }
 
-        this.state[pacboyY][pacboyX] = 'C';
+        deletePacboy();
+        pacboyY = desiredY;
+
+        this.state[pacboyY][pacboyX] = CHAR_PLAYER;
     }
 
     public void desce() {
+        int desiredY = Math.abs((pacboyY + 1) % this.size);
+        if (shouldBeABlock(pacboyX, desiredY)) {
+            return;
+        }
         deletePacboy();
-        pacboyY = Math.abs((pacboyY + 1) % this.size);
+        pacboyY = desiredY;
 
-        this.state[pacboyY][pacboyX] = 'C';
+        this.state[pacboyY][pacboyX] = CHAR_PLAYER;
     }
 
     public void tick() {
